pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["orji"]

[project]
name = "orji"
authors = [
    {name = "Colm O'Connor", email = "colm.oconnor.github@gmail.com"},
]
description = "Org mode to jinja2 templating."
license = {text = "MIT"}
requires-python = ">=3.7.0"
keywords = ["orgmode", "markdown", "latex", "jinja2"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Text Processing :: Markup",
    "Topic :: Software Development :: Libraries",
    "Natural Language :: English",
]
dependencies = [
    "git+https://github.com/durableOne/orgmunge.git@dev",
    "orgparse>=0.3.1",
    "jinja2>=3.1.2",
    "click>=8.1.3",
    "python-slugify>=7.0.0",
]
dynamic = ["version", "readme"]

[project.scripts]
orji = "orji:cli"

[project.urls]
homepage = "https://hitchdev.com/orji"
documentation = "https://hitchdev.com/orji/using"
repository = "https://github.com/crdoconnor/orji"
changelog = "https://hitchdev.com/orji/changelog"

[tool.setuptools.dynamic]
readme = {file = ["README.md",], content-type = "text/markdown"}
version = {file = "VERSION"}
